[#24]refactor: api type들 파일 분리
