+++ client/src/lib/api/types/api.ts
@@ -1,0 +1,5 @@
+interface ApiResponse {
+  statusCode: number;
+}
+
+export default ApiResponse;
