+++ client/src/lib/api/types/auth.ts
@@ -1,0 +1,7 @@
+import ApiResponse from './api';
+
+export interface AuthResponseBody {
+  access: string;
+}
+
+export interface AuthResponse extends ApiResponse, AuthResponseBody {}
