+++ client/src/lib/api/types/index.ts
@@ -1,18 +1,3 @@
-export interface ApiResponse {
-  statusCode: number;
-}
-
-export interface ErrorResponseBody {
-  success: boolean;
-  statusCode: number;
-  message: string;
-  stack?: string;
-}
-
-export interface AuthResponseBody {
-  access: string;
-}
-
-export interface ErrorResponse extends ApiResponse, ErrorResponseBody {}
-
-export interface AuthResponse extends ApiResponse, AuthResponseBody {}
+export { default as ApiResponse } from './api';
+export * from './error';
+export * from './auth';
